requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2
django-filter==25.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.12
psycopg-binary==3.2.12
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
psycopg2-binary
